+++ src/reading.rs
@@ -11,9 +11,8 @@ where
     ///
     /// Indicates that the RGBC channels have completed an integration cycle.
     pub fn is_rgbc_status_valid(&mut self) -> Result<bool, Error<E>> {
-        let mut status = [0];
-        self.read_register(Register::STATUS, &mut status)?;
-        Ok((status[0] & BitFlags::RGBC_VALID) != 0)
+        let status = self.read_register(Register::STATUS)?;
+        Ok((status & BitFlags::RGBC_VALID) != 0)
     }
 
     /// Read the clear (unfiltered) channel measurement data.
@@ -36,14 +35,32 @@ where
         self.read_channel(Register::BDATA)
     }
 
-    fn read_channel(&mut self, register: u8) -> Result<u16, Error<E>> {
+    fn read_channel(&mut self, first_register: u8) -> Result<u16, Error<E>> {
         let mut cdata = [0; 2];
-        self.read_register(register, &mut cdata)?;
+        self.read_registers(first_register, &mut cdata)?;
         Ok((cdata[1] as u16) << 8 | cdata[0] as u16)
     }
 
-    fn read_register(&mut self, register: u8, mut data: &mut [u8]) -> Result<(), Error<E>> {
-        let command = BitFlags::CMD | BitFlags::CMD_AUTO_INC | register;
+    /// Read the device ID.
+    ///
+    /// The value returned corresponds to the part number identification:
+    /// - `0x44` => `TCS34725`
+    /// - `0x4D` => `TCS34727`
+    pub fn read_device_id(&mut self) -> Result<u8, Error<E>> {
+        self.read_register(Register::ID)
+    }
+
+    fn read_register(&mut self, register: u8) -> Result<u8, Error<E>> {
+        let command = BitFlags::CMD | register;
+        let mut data = [0]; 
+        self.i2c
+            .write_read(DEVICE_ADDRESS, &[command], &mut data)
+            .map_err(Error::I2C)?;
+        Ok(data[0])
+    }
+
+    fn read_registers(&mut self, first_register: u8, mut data: &mut [u8]) -> Result<(), Error<E>> {
+        let command = BitFlags::CMD | BitFlags::CMD_AUTO_INC | first_register;
         self.i2c
             .write_read(DEVICE_ADDRESS, &[command], &mut data)
             .map_err(Error::I2C)